move python extraction rules behind the indexer language registry
S3-PYTHON (ARCH-012), the sibling of S3-KOTLIN and the second one-language
rollback unit to land. Python's tree-sitter rule file, compiled-rule cache,
attribute-call callsite marker, decorator-edge collector, SCREAMING_CASE
constant predicate, and receiver-call resolution engine move out of the
27.4k-line indexer god file into `codestory-indexer/src/languages/python.rs`,
behind the existing `LanguageExtraction` registry in `languages/mod.rs`. Every
language-keyed dispatch that used to spell "python" now reads the registry
first and keeps its residual `match` for the fourteen languages that have not
moved; the last sibling package deletes those arms.

The move is mechanical. 44 free functions left `lib.rs` (673 -> 629 top-level
`fn`) and the same 44 landed in `languages/python.rs` alongside the new
`python_language` grammar constructor. With whitespace collapsed, the moved
bodies are character-identical to the deleted lines (21,256 chars each side)
once six declared boundary renames are inverted:
`collect_python_receiver_call_edges` -> `receiver_call_specs`,
`python_member_call` -> `member_call`,
`python_attribute_method_col` -> `attribute_method_col`,
`is_python_implicit_receiver` -> `is_implicit_receiver`,
`is_python_constant_name` -> `is_constant_name`,
`collect_python_decorator_call_edges` -> `decorator_call_specs`, plus
`PYTHON_ATTRIBUTE_CALLSITE_MARKER` -> `MEMBER_CALLSITE_MARKER`. The only
line-level difference is rustfmt breaking `attribute_method_col`'s signature
across four lines, because `pub(crate) fn ` pushed it past the width limit.

Five of those entry points stay `pub(crate)` because their call sites are
shared seams with no registry field: the manual-edge collector, the shared
`member_call_method_col`, the shared receiver-edge appender, the graph-node
projection, and the two placeholder-annotation rosters. Each seam keeps its
`language_name == "python"` guard in `lib.rs` and calls into the module for
the content. `collect_python_route` stays in `lib.rs` for the same reason
Kotlin's `collect_ktor_route` did: the per-language route collectors take
non-uniform arguments, so routing them through the registry is one change for
all sixteen languages rather than part of Python's rollback unit.

Proof: `tests/language_extraction_snapshot.rs` gains a Python row that pins the
full rendered projection of both Python fixtures -- the fidelity_lab file
through the store pipeline (resolution included, 57 nodes / 37 edges, five of
them carrying `syntax:python-attribute-call`) and the tictactoe file through
raw `index_file` (161 nodes / 125 edges). Both goldens were captured before the
production edit and are byte-identical after it (sha256 6f5345ef.. /
501f11f4..). The threshold suites are not sufficient on their own: changing
Python's qualified-name delimiter to `::` leaves both `fidelity_regression` and
`tictactoe_language_coverage` fully green while the new snapshots go red on
`ConsoleNotifier.notify` becoming `ConsoleNotifier::notify`, and flipping
`promotes_type_member_functions_to_methods` leaves `fidelity_regression` green.

Declared behavior changes:
- None to any indexing, resolution, or rendering output. Python
  fidelity_regression and tictactoe_language_coverage projections are proven
  byte-identical before and after, and packet_search_eval is unmoved.
- New crate-internal module `codestory-indexer/src/languages/python` (private
  `mod` under a private `mod`, no public API change to the indexer crate).
- `LanguageRuleset::Python`'s arm in `compiled_rules` is now unreachable and
  returns a typed error instead of compiled rules; the registry answers first.
  Fail-closed by construction, matching the Kotlin arm.
- `codestory_contracts::language_support::LANGUAGE_COMMENT_PROFILES` gains a
  `python` -> `#` row and the CLI's `ansi_highlight_line` roster loses its
  `python` arm. The rendered marker is unchanged; the roster's own
  restated-table test proves every other language is unmoved. The CLI comment
  roster and the indexer route-comment roster still disagree deliberately and
  are left alone.
- `LANGUAGE_SUPPORT_PROFILES` is unchanged, so extension routing, claim tiers,
  `supported_extensions`, `is_structural_source_path`, and every existing
  language claim are identical.
- No version surface touched; everything stays 0.16.3.

Closes #1683
Refs #1243
Refs #1179
